pyproject.toml
[tool.poetry]
name = "first-django-app"
version = "0.1.0"
description = ""
authors = ["Sn1kls <macks.chigrin@gmail.com>"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django-extensions = "^3.2.3"
django = "^5.1.2"
pydotplus = "^2.0.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^23.0.0"
redis = "^5.2.0"
django-redis = "^5.4.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
celery = "^5.4.0"
drf-spectacular = "^0.27.2"
django-cors-headers = "^4.6.0"
django-filter = "^24.3"
django-ratelimit = "^4.1.0"
method-decorator = "^0.1.3"
daphne = "^4.1.2"
channels = {extras = ["daphne"], version = "^4.2.0"}

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", ".venv", ".vscode", ".vscode-test", ".pytest", "migrations", "settings.py", "manage.py", "db.sqlite3", ".mypy_cache", ".pytest_cache", ".vscode", ".vscode-test", ".gitignore", ".flake8", ".editorconfig", ".env", ".env.example", ".dockerignore", "Dockerfile"]

[tool.isort]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.black]
line-length = 120
target-version = ["py312"]
exclude = """
(
    | .+/migrations
    | .+/venv
    | .+/tests
)
"""